net/iavf: fix RSS configuration failure
According to patch [1], IPv4 and IPv6 fragment ptypes should be
separated from IP other ptypes. That is to say, when RSS for IP
packets is enabled, it should not affect IP fragment packets.

Previously, we try to enable RSS for both IP packets and IP fragment
packets with hash field of src and dst address. However, due to
share code limitation, when RSS for IP packets and IP fragment packets
coexists, they cannot share the same hash field.

As a result, independent configuration of default RSS for IP fragment
packets need to be removed.

This patch revert the original patch to fix this failure.

[1]
commit 08c16b212d03 ("net/ice/base: fix ptype bitmap for IP fragment")

Fixes: c40525568480 ("net/iavf: fix default RSS hash for IP fragment packets")
Fixes: 9e29a278bc0c ("net/iavf: support default RSS for IP fragment")
Cc: stable@dpdk.org

Signed-off-by: Wenjun Wu <wenjun1.wu@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>