app/testpmd: update forward engine beginning
For each forward engine, there may be some special conditions
must be meet before the forwarding run.

Adding checks for these conditions in configuring is not suitable,
because one condition may rely on multiple configurations, and the
conditions required by each forward engine is not general.

The best solution is each forward engine has a callback to check
whether these conditions are meet, and then testpmd can call the
callback to determine whether the forwarding can be started.

There was a void callback 'port_fwd_begin' in forward engine,
it did some initialization for forwarding, this patch updates it's
return type, then we can add some checks in it to confirm whether
the forwarding can be started. In addition, this patch puts the
calling part of the callback up to before some forwarding related
status being set.

Signed-off-by: Alvin Zhang <alvinx.zhang@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>