net/hns3: fix residual old MAC after setting default MAC
This problem occurs in the following scenarios:
1) reset is encountered when the adapter is running.
2) set a new default MAC address

After the above two steps, the old default MAC address should be not take
effect. But the current behavior is contrary to that. This is due to the
change of the "default_addr_setted" in hw->mac from 'true' to 'false' after
the reset. As a result, the old MAC address is not removed when the new
default MAC address is set. This variable controls whether to delete the
old default MAC address when setting the default MAC address. It is only
used when the mac_addr_set API is called for the first time. In fact, when
a unicast MAC address is deleted, if the address isn't in the MAC address
table, the driver doesn't return failure. So this patch remove the
redundant and troublesome variables to resolve this problem.

Fixes: 7d7f9f80bbfb ("net/hns3: support MAC address related operations")
Cc: stable@dpdk.org

Signed-off-by: Huisong Li <lihuisong@huawei.com>
Signed-off-by: Min Hu (Connor) <humin29@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>