examples/vhost: fix use-after-free on drain vhost
When a vdev is removed in destroy_device function,
the corresponding vhost_txbuff[vdev->vid] will also
be free, but the vhost_txbuff[vdev->vid] may still
be used in the drain_vhost function, which will cause
an error of heap-use-after-free. Therefore, before
accessing vhost_txbuff[vdev->vid], we need to check
whether the vdev has been removed, if so, let's skip
this vdev.

Signed-off-by: Wenwu Ma <wenwux.ma@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>