net/mlx5: fix xstats get reinitialization
The mlx5_xstats_get gets the device extended statistics.
In this function the driver may reinitialize the structures that are
used to read device counters.

In case of reinitialization, the number of counters may change, which
wouldn't be taken into account by the get API callback and can cause
a segmentation fault.

In case of insufficient supplied stats table size, ex. zero to query
the number of extended stats, reinitialization may never happen and
the returned stats number, that is used for subsequent stats getting,
will not be sufficient too.

This issue is fixed by getting and allocating the counters size after
the reinitialization.

Fixes: 1a611fdaf6ec ("net/mlx5: support missing counter in extended statistics")
Fixes: a4193ae3bc4f ("net/mlx5: support extended statistics")
Cc: stable@dpdk.org

Signed-off-by: Vladislav Grishenko <themiron@yandex-team.ru>
Signed-off-by: 0-day Robot <robot@bytheb.org>