telemetry: use unique socket paths for in-memory mode
When DPDK is run using "in-memory" flag, multiple processes can be run
using the same file-prefix and hence the same runtime directory. To
avoid problems with conflicting telemetry unix socket paths, we can
put the pid of the process into the socket name. As with the existing
telemetry socket files, these sockets are removed on normal program
exit.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>