app/flow-perf: add destination ports parameter
Add destination ports(dst-ports) parameter for port-id action,
the parameter is only valid for port-id action. The parameter is not
Must, and the value is 1 by default as before if not provided.

For example:

$ dpdk-test-flow-perf -w 08:00.0,representor=[0,1] -- --transfer \
> --ingress --transfer --ether --portmask=0x2 --vxlan-encap \
> --port-id --dst-ports=0

This command means the rule created on representor 0 with port 0
as destination, since the portmask is 0x2 and dst-ports is 0:

$ dpdk-test-flow-perf -w 08:00.0,representor=[0,1] \
> -w 08:00.1,representor=[0,1]-- --transfer --ingress --transfer \
> --ether --portmask=0x12 --vxlan-encap --port-id --dst-ports=0,3

This command means the rules created on both representor 0 of PF 0
and PF 1, the destination port for the first represontor is PF 0,
and the destination port for the other one it PF 1.

Signed-off-by: Sean Zhang <xiazhang@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>